Update weight functionality

+++ app/src/main/java/com/github/anastasiazhukova/fitness/screens/fragments/register/presentation/RegisterFragment.kt
@@ -133,7 +133,7 @@ class RegisterFragment : Fragment(R.layout.fragment_register) {
             trainerNickname = trainerNickname.getTextAsString(),
             gender = getGenderValue(),
             height = height.asInt(),
-            weight = weight.asInt(),
+            weight = weight.asFloat(),
             fitnessLevel = getFitnessLevelValue(),
             goalCalories = goalCalories.asInt(),
             goalWater = goalWater.asInt()
