+++ fitnessadmin/src/main/java/com/github/anastasiazhukova/fitness/admin/screens/fragment/clients/domain/ClientModel.kt
@@ -8,7 +8,7 @@ data class ClientModel(
     val trainerNickname: String,
     val gender: Boolean,
     val height: Int?,
-    val weight: Int?,
+    val weight: Float?,
     val fitnessLevel: Int?,
     val goalCalories: Int?,
     val goalWater: Int?,
